# docs/integrations/reference/zypto-fcfpay-openapi.yaml
openapi: 3.0.3
info:
  title: FCFpay
  description: ''
  version: 1.0.0
servers:
  -
    url: 'https://merchant.fcfpay.com'
paths:
...
  /api/cards/issue-card-api:
    post:
      summary: 'Issue Card (Virtual)'
      operationId: issueCardVirtual
      description: "This API call will generate a Prepaid virtual card and load it with the amount specified in the \"amount\" field.\n\nFor this API call to be successfull, you will need to have \"Allowance mode\" activated and have sufficient allowance funds.\n\nIf card_type is not sent, default value 1 will be assumed."
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer 5ScRiAOgI5BlgdjSwWNUbNagsj8fFCjady'
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: "{\n     'success': true,\n     'message': {\n          \"card_info\": {\n             \"status\": \"Approved\",\n             \"card_number\": \"4601380590023942\",\n             \"card_exp_mth\": \"11\",\n             \"card_exp_year\": \"2023\",\n             \"card_cvv\": \"128\"\n          },\n         \"card_balance\": {\n             \"status\": \"Approved\",\n             \"balance\": 50\n         }\n         'card_id' => 1\n     },\n}"
        401:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: "{\n     'success': false,\n     'message' => 'Something went wrong!',\n}"
      tags:
        - 'Virtual cards api'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                iframe_id:
                  type: integer
                  description: ''
                  example: 14
                amount:
                  type: number
                  description: 'Must match the regex /^\d+(\.\d{1,2})?$/. Must be at least 20.'
                  example: 82
                currency:
                  type: string
                  description: 'Must be at least 3 characters.'
                  example: sadcrcozwepmzxigwhwr
                product_id:
                  type: integer
                  description: ''
                  example: 7
                email:
                  type: string
                  description: 'Must be a valid email address. Must match the regex /^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/. Must be at least 3 characters.'
                  example: maybell13@example.net
              required:
                - iframe_id
                - amount
                - currency
                - email
      security: []
...
  /api/cards/get-card-balance:
    post:
      summary: 'Get Card Balance'
      operationId: getCardBalance
      description: 'This API call will return the current balance of the specified card.'
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer 5ScRiAOgI5BlgdjSwWNUbNagsj8fFCjady'
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: "{\n     'success': true,\n     'data': {\n         \"status\": \"Approved\",\n         \"balance\": 1000\n     }\n}"
        401:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: "{\n     'success': false,\n     'message' => 'Something went wrong!',\n}"
      tags:
        - 'Virtual cards api'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                card_number:
                  type: string
                  description: 'Must not be greater than 20 characters. Must be at least 3 characters.'
                  example: donccmuqxwyoyelocwnqfadobyxdsdncwtxtesaulnuytd
              required:
                - card_number
      security: []
...
  /api/cards/load-virtual-card:
    post:
      summary: 'Load Virtual Card'
      operationId: loadVirtualCard
      description: 'This call will attempt to load the card with the amount specified.'
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer 5ScRiAOgI5BlgdjSwWNUbNagsj8fFCjady'
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: "{\n     'success': true,\n     'message' => status message,\n     'load_amount' => 100,\n     'allowance_used' => 0\n}"
        401:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: "{\n     'success': false,\n     'message' => 'Something went wrong!',\n}"
      tags:
        - 'Virtual cards api'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                card_number:
                  type: string
                  description: 'Must not be greater than 20 characters. Must be at least 3 characters.'
                  example: vqwnmwfsffoavnsawhspprnhorratetivshckfalsm
                amount:
                  type: number
                  description: 'Must match the regex /^\d+(\.\d{1,2})?$/. Must be at least 20.'
                  example: 41
              required:
                - card_number
                - amount
      security: []
...
  /api/cards/get-my-cards:
    post:
      summary: 'Get My Cards'
      operationId: getMyCards
      description: ''
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer 5ScRiAOgI5BlgdjSwWNUbNagsj8fFCjady'
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: "{\n     'success': true,\n     'cards': [{\n         'holder_name' => $holder_name,\n         'card_number' => $card_number,\n         'expiry_date' => $expiry_date,\n         'cvc' => 111,\n         'balance' => 100\n     }]\n}"
        401:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: "{\n     'success': false,\n     'message' => 'Something went wrong!',\n}"
      tags:
        - 'Virtual cards api'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                cards_type:
                  type: string
                  description: ''
                  example: atque
              required:
                - cards_type
      security: []
...
  /api/general/available_currencies:
    get:
      summary: 'Available Currencies'
      operationId: availableCurrencies
      description: ''
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer 5ScRiAOgI5BlgdjSwWNUbNagsj8fFCjady'
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: "{\n     'success': true,\n     'data': {\n         \"ticker\" => 'BTC',\n         \"token_symbol\" => 'BTC.BTC',\n         \"name\" => 'BTC',\n         \"blockchain\" => 'Bitcoin',\n         \"block_explorer\" => 'https://www.blockchain.com/btc/tx',\n         \"image\" => image link,\n         \"isStable\" => 1/0,\n         \"enabled_on_account\" => true/false\n }\n}"
        401:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: "{\n     'success': false,\n     'message' => 'Something went wrong!',\n}"
      tags:
        - 'Virtual cards api'
      security: []
...